+++ NEWS.md
@@ -1,3 +1,9 @@
+# cyclotomic 1.3.0
+
+The package does no longer depend on the 'primes' package; instead, it now 
+depends on the 'numbers' package.
+
+
 # cyclotomic 1.2.0
 
 Function `quadraticRoots`, returning the roots of a polynomial of degree 2 with rational coefficients.